Add missing build flags and extra optimization training rules
Turns out that other distros are using -DUNALIGNED_OK -D_REENTRANT and
-D_LARGEFILE64_SOURCE=1. The -DUNALIGNED_OK allows processors that
can do unaligned load/stores to perform unaligned data access, instead
having to add the overhead to cater for the misalignment. This is quite
a win. Also add the -D_REENTRANT and -D_LARGEFILE64_SOURCE=1 flags, no
performance loss and keeps it in-sync with other distros.

Use a binary executable as part of the training data to cater for
compression of the more typical non-ASCII based binary data, e.g.
compressing tarballs of object code.

This was benchmarked using 29 different kinds of input streams from
the stress-ng zlib stress case (that uses zlib) on an i9-12900.
The geometric mean of the compression rates (in MB/s) results was
taken and are as follows:

Clear Linux (Pre-change):   14.84
Clear Linux (Post-change):  17.27
Ubuntu 23.04:               16.88 (as a comparison)

So, quite a big win.

Signed-off-by: Colin Ian King <colin.king@intel.com>